Add band constraints to steps with faro tasks.
When a task has band in its quantum dimensions but does not have band
in the dimensions of any of its inputs, it will run over all bands
known to any instrument unless those bands are constrained by the data
query.  It was an accident of the old QG generation algorithm that
this wasn't happening before (unrelated tasks in the same step were
providing a band constraint, but the middleware now sees them as
unrelated).